Here is my submission for Hacktoberfest!
Thanks for this awesome experience, unfortunately, i haven't the time that i wanted to work on this project. I've becomed dad recently thats why i didn't have the same time as before. So i send my request now as we are the 30 and i will not be able to work anymore on the project before the end of Hacktoberfest!

I hope it will be merged, if not, thanks anyway for the knowledge and the experience gained in this event! I will definitively come back to it later in November, and look for the disney clone that look cool too!

Thanks again and have a nice day!

V0lk3n